try again

+++ src/CloudTasksJob.php
@@ -105,7 +105,9 @@ class CloudTasksJob extends LaravelJob implements JobContract
     {
         parent::delete();
 
-        $this->cloudTasksQueue->delete($this);
+        if ($this->job['internal']['attempts'] === $this->maxTries) {
+            $this->cloudTasksQueue->delete($this);
+        }
     }
 
     public function release($delay = 0)
